Clean up codebase

+++ browserup-proxy-core/src/test/java/com/browserup/harreader/model/HarTimingTest.java
@@ -11,13 +11,13 @@ public class HarTimingTest extends AbstractMapperTest<HarTiming> {
         + "\"receive\": 79, \"ssl\": 123, \"comment\": \"my comment\"}", HarTiming.class);
 
         Assert.assertNotNull(timing);
-        Assert.assertEquals(3804, (int) timing.getBlocked());
-        Assert.assertEquals(23, (int) timing.getDns());
-        Assert.assertEquals(5, (int) timing.getConnect());
-        Assert.assertEquals(9, (int) timing.getSend());
-        Assert.assertEquals(5209, (int) timing.getWait());
-        Assert.assertEquals(79, (int) timing.getReceive());
-        Assert.assertEquals(123, (int) timing.getSsl());
+        Assert.assertEquals(3804, timing.getBlocked());
+        Assert.assertEquals(23, timing.getDns());
+        Assert.assertEquals(5, timing.getConnect());
+        Assert.assertEquals(9, timing.getSend());
+        Assert.assertEquals(5209, timing.getWait());
+        Assert.assertEquals(79, timing.getReceive());
+        Assert.assertEquals(123, timing.getSsl());
         Assert.assertEquals("my comment", timing.getComment());
     }
 
@@ -25,27 +25,27 @@ public class HarTimingTest extends AbstractMapperTest<HarTiming> {
     public void testBlocked() {
         HarTiming timing = new HarTiming();
         timing.setBlocked(null);
-        Assert.assertEquals(-1, (int) timing.getBlocked());
+        Assert.assertEquals(-1, timing.getBlocked());
     }
 
     @Test
     public void testDns() {
         HarTiming timing = new HarTiming();
         timing.setDns(null);
-        Assert.assertEquals(-1, (int) timing.getDns());
+        Assert.assertEquals(-1, timing.getDns());
     }
 
     @Test
     public void testConnect() {
         HarTiming timing = new HarTiming();
         timing.setConnect(null);
-        Assert.assertEquals(-1, (int) timing.getConnect());
+        Assert.assertEquals(-1, timing.getConnect());
     }
 
     @Test
     public void testSsl() {
         HarTiming timing = new HarTiming();
         timing.setSsl(null);
-        Assert.assertEquals(-1, (int) timing.getSsl());
+        Assert.assertEquals(-1, timing.getSsl());
     }
 }
